Make sure that Condition objects which have instancemethod arguments are pickleable.
Turns out instancemethod types are no pickleable.  To work around this limitation, when an argument is provided to the Condition constructor, it is broken apart into its module, class and func parts.  Those are then stored for later use.  An argument @property is defined, and when requested, rebuilds the argument based on those parts.  The resulting argument is cached locally in the instance of the Condition.