Fixed issue where global inputs could get cached for parent switches.
This is a workaround for a bug where an input could be unpicklable, and
auto-creating switches would fail when the manager has an unpicklable
input.

The underlying issue (which still exists) is that Manager.register handles
unsetting Switch.manager before persisting except it does not take into
account parent switches having managers attached still.