Encode switch names into bytes
Background idea is that we use string types as switch names and if we do
this consistently, then the keys can be encoded into bytes objects.

At least the RedisDict as storage will otherwise happily take str
instances and return bytes instances. Result is that the Manager cannot
find the switches anymore and we see issues due to mixing up str and
bytes.